requirements.txt
paho-mqtt==2.1.0
PyMySQL==1.1.2
python-dotenv==1.2.1
RPi.GPIO==0.7.1
